sqltoolkit_mcp/requirements.txt
fastmcp
python-dotenv
requests
azure-identity
pyodbc
psycopg2-binary
snowflake-connector-python
